Takes screenshot and saves it in a folder
Every 15 seconds, the program asks if a screenshot is required. If yes, within the next 5 seconds, it takes the screenshot, giving the user time to set the window to be captured. This loop continues forever until the user terminates the process. I got this idea when I was working with PostgreSQL commands for practice and wanted to capture the commands and their respective output in a folder that I could later push to GitHub.